chore: rename loader to document_loader, move to apps
